+++ docs/examples/python/example/resolvers.py
@@ -1,0 +1,4 @@
+from reactpy_router.resolvers import ReactPyResolver
+
+
+class CustomResolver(ReactPyResolver): ...